+++ app/public/model/PrimaryModel.js
@@ -48,11 +48,12 @@ export default class PrimaryModel extends Observable {
         this.notify();
     }
 
-    handleLocationChange() {
+    async handleLocationChange() {
         const url = this.router.getUrl();
         switch (url.pathname) {
-            case '/':
-                if (! this.router.params['page']) {
+            case '/': {
+                const { page } = this.router.params;
+                if (! page) {
                     this.router.go(`/?page=${pageNames.periods}&${dataReqParams.rowsOnSite}=50&${dataReqParams.site}=1&sorting=-name`);
                 } else {
                     this.fetchedData.reqForData()
@@ -60,11 +61,58 @@ export default class PrimaryModel extends Observable {
                         .catch(() => {});
                 }
                 break;
+            }
             case '/admin/':
                 throw 'TODO';
             default:
                 break;
         }
+
+        /*
+         *
+         *Switch (url.pathname) {
+         *    case '/':
+         *        const page = this.router.params['page'];
+         *        if (! page) {
+         *            this.router.go(`/?page=${pageNames.periods}&${dataReqParams.rowsOnSite}=50&${dataReqParams.site}=1&sorting=-name`);
+         *        } else {
+         *            switch (page) {
+         *                case pageNames.flags:
+         *                    const dataPassName = this.router.params['data_pass_name'];
+         *                    console.log(dataPassName);
+         *                    if (dataPassName) {
+         *                        /*
+         *                        console.log(dataPassName);
+         *                        await this.parent.runs.fetchRunsPerDataPass(dataPassName).then(() => {}).catch((e) => {console.log(e)});
+         *                        console.log(this.fetchedData);
+         *                        // console.log(this.fetchedData);
+         *
+         *                        const dpSearchParams = `?page=${pageNames.runsPerDataPass}&index=${dataPassName}&${dataReqParams.rowsOnSite}=50&${dataReqParams.site}=1`;
+         *                        const dpUrl = new URL(url.origin + url.pathname + dpSearchParams);
+         *                        this.fetchedData.reqForData(true, dpUrl).then(() => {
+         *                            const runNumbers = this.fetchedData[pageNames.runsPerDataPass][dataPassName].payload.rows.map((row) => row.run_number);
+         *                            this.parent.runs.fetchFlagsSummary(dataPassName, runNumbers).then(() => {
+         *                                this.fetchedData.reqForData();
+         *                            }).catch(() => {});
+         *                        });
+         *
+         *                    } else this.goToDefaultPageUrl(pageNames.flags);
+         *                    break;
+         *                default:
+         *                    this.fetchedData.reqForData()
+         *                        .then(() => {})
+         *                        .catch(() => {});
+         *                    break;
+         *            }
+         *        }
+         *        break;
+         *    case '/admin/':
+         *        throw 'TODO';
+         *    default:
+         *        break;
+         *}
+         *
+         */
     }
 
     async logout() {
